refactor: restore local storage constants for access and refresh tokens
